Refactored helper functions in ws.go
